+++ packages/fhir-eswatini/types/Adaptor.d.ts
@@ -1,4 +1,5 @@
 export { read, search, update, delete, create, createBundle, uploadBundle, } from '@openfn/language-fhir-4';
+export declare const BASE_URL = "https://hapifhir.eswatinihie.com";
 /**
  * Add a resource to a bundle on state, using the `name` as the key (or `bundle` by default).
  * The resource will be upserted (via PUT).